+++ src/MuMu_RichPresence/Tray/RichPresence_Tray.cs
@@ -9,7 +9,7 @@ using WinForms.ContextMenu;
 
 public class RichPresence_Tray
 {
-    internal NotifyIcon Tray { get; private set; }
+    internal NotifyIcon Tray { get; }
     private readonly ILogger _logger = Log.ForContext<RichPresence_Tray>();
     public RichPresence_Tray()
     {
@@ -62,7 +62,7 @@ public class RichPresence_Tray
         }
     }
 
-    private void LogInteractionsRecursively(ToolStripItemCollection items)
+    private static void LogInteractionsRecursively(ToolStripItemCollection items)
     {
         foreach (ToolStripItem item in items)
         {
@@ -98,9 +98,9 @@ public class RichPresence_Tray
         return enabledItem;
     }
 
-    private void ChangeEnabledStateOnStartupIfNecessary(bool enabled)
+    private static void ChangeEnabledStateOnStartupIfNecessary(bool enabled)
     {
-        if (!Startup.StartsWithWindows(Application.ProductName!))
+        if (!Startup.StartsWithWindows(Application.ProductName!, Application.ExecutablePath))
             return;
 
         Startup.StartWithWindows(Application.ProductName!,
@@ -112,10 +112,10 @@ public class RichPresence_Tray
 
     private ToolStripMenuItem HideTray() => new("Hide Tray", null, (_, _) => Tray.Visible = false);
 
-    private ToolStripMenuItem RunOnStartup()
+    private static ToolStripMenuItem RunOnStartup()
     {
         var startup = new ToolStripMenuItem("Run on Startup");
-        startup.Checked = Startup.StartsWithWindows(Application.ProductName!);
+        startup.Checked = Startup.StartsWithWindows(Application.ProductName!, Application.ExecutablePath);
 
         startup.Click += delegate
         {
@@ -132,7 +132,7 @@ public class RichPresence_Tray
 
     private ToolStripMenuItem Exit() => new("Exit", null, (_, _) => Tray.Dispose());
 
-    private ToolStripItem[] Header()
+    private static ToolStripItem[] Header()
     {
         var header = new ToolStripMenuItem("MuMu Player Rich Presence");
         header.Enabled = false;